move trigger_rate_limit to DefaultTriager
DefaultTriager.start might call it
